fast/requirements.txt
fastapi==0.115.5
psycopg2-binary== 2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.41.3
typing_extensions
uvicorn==0.23.2
annotated-types==0.6.0